move modern TextWithCopy
