+++ src/screens/WalletMigrationArchanova/SelectAssets/WalletSummary.js
@@ -24,7 +24,7 @@ import { useTranslationWithPrefix } from 'translations/translate';
 // Components
 import BalanceView from 'components/BalanceView';
 import Text from 'components/core/Text';
-import TextWithCopy from 'components/other/TextWithCopy';
+import TextWithCopy from 'components/display/TextWithCopy';
 
 // Utils
 import { useThemeColors } from 'utils/themes';
